# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/generos": {
      "post": {
        "operationId": "GeneroController_crear",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearGeneroDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Genero"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos: el mensaje indica qué campo falló, un id de la URL que no es un número, o un id del body que no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión: falta el token, es inválido o está vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso: la acción es solo para ADMIN o el recurso es de otro usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto: el dato ya existe (duplicado) o está en uso y no se puede borrar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear un género (solo ADMIN)",
        "tags": [
          "Géneros"
        ]
      },
      "get": {
        "operationId": "GeneroController_listar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Genero"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar géneros (público)",
        "tags": [
          "Géneros"
        ]
      }
    },
    "/generos/{id}": {
      "get": {
        "operationId": "GeneroController_buscarPorId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Genero"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos: el mensaje indica qué campo falló, un id de la URL que no es un número, o un id del body que no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "404": {
            "description": "No existe el recurso pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ver un género por id (público)",
        "tags": [
          "Géneros"
        ]
      },
      "patch": {
        "operationId": "GeneroController_actualizar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarGeneroDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Genero"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos: el mensaje indica qué campo falló, un id de la URL que no es un número, o un id del body que no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión: falta el token, es inválido o está vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso: la acción es solo para ADMIN o el recurso es de otro usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "404": {
            "description": "No existe el recurso pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto: el dato ya existe (duplicado) o está en uso y no se puede borrar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Modificar un género (solo ADMIN)",
        "tags": [
          "Géneros"
        ]
      },
      "delete": {
        "operationId": "GeneroController_eliminar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Datos inválidos: el mensaje indica qué campo falló, un id de la URL que no es un número, o un id del body que no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión: falta el token, es inválido o está vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso: la acción es solo para ADMIN o el recurso es de otro usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "404": {
            "description": "No existe el recurso pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto: el dato ya existe (duplicado) o está en uso y no se puede borrar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar un género (solo ADMIN; 409 si algún juego lo usa)",
        "tags": [
          "Géneros"
        ]
      }
    },
...
  },
  "info": {
    "title": "DGame (MyGameSearcher) API",
    "description": "API REST que recomienda de 1 a 3 videojuegos según géneros, características y plataforma.\n\n**Cómo autenticarse:** hacer `POST /auth/login` con email y contraseña, copiar el `token` de la respuesta y pegarlo en el botón **Authorize** (arriba a la derecha). Desde ahí, todas las pruebas se mandan con `Authorization: Bearer <token>`.\n\n**Niveles de acceso:** público (sin token), usuario con sesión iniciada, y ADMIN. El resumen de cada ruta indica quién puede usarla.\n\n**Errores:** todas las respuestas de error tienen la forma `{ \"statusCode\", \"message\", \"error\" }`, con el mensaje en español. 400: datos inválidos · 401: sin sesión o token inválido · 403: sin permiso · 404: no existe · 409: conflicto (duplicado o en uso).",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Token que devuelve POST /auth/login"
      }
    },
    "schemas": {
      "CrearGeneroDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 50
          }
        },
        "required": [
          "nombre"
        ]
      },
      "Genero": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "nombre": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "nombre"
        ]
      },
      "ErrorRespuestaDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "message": {
            "description": "Mensaje en español. En los errores de validación (400) es una lista con un mensaje por campo",
            "oneOf": [
              {
                "type": "string",
                "example": "No se encontró el género con id 9999"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "El nombre es obligatorio"
                ]
              }
            ]
          },
          "error": {
            "type": "string",
            "example": "Not Found"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
      "ActualizarGeneroDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 50
          }
        }
      },
...
    }
  }
}